+++ packages/react/src/locale/index.ts
@@ -1,0 +1,8 @@
+export type {
+  EkycLocale,
+  EkycPreviewLocale,
+  EkycRoomLocale,
+  EkycPanelLocale,
+} from "./types";
+export { viLocale } from "./vi";
+export { enLocale } from "./en";